+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Auto suggest filename based on texture name
+
 ## [1.1.0] - 2021-04-27
 
 ### Fixed